backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
cors==1.0.1
numpy
yfinance
feedparser
sqlalchemy>=2.0
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
python-multipart
